evals/runner/requirements.txt
httpx>=0.25.0
pyyaml>=6.0
jsonschema>=4.20.0
pydantic>=2.0.0
click>=8.1.0
rich>=13.0.0
python-Levenshtein>=0.23.0
